- Fix event speaker getting sprite provider incorrectly - Safely cache all images on load so the files are accessible when patching

+++ RanseiLink/ViewModels/ModelViewModels/GimmickViewModel.cs
@@ -185,7 +185,7 @@ public class GimmickViewModel : GimmickViewModelBase
         }
     }
 
-    public string ImagePath => _spriteProvider.GetSpriteFilePath(SpriteType.StlStageObje, (uint)Image);
+    public string ImagePath => _spriteProvider.GetSpriteFilePath(SpriteType.StlStageObje, Image);
 }
 
 public class GimmickGridItemViewModel : GimmickViewModelBase
